template: improve the confidence section
Currently, it asks:

> How sure are we that this impact will be realized?

Where "impact" is:

> How would this directly contribute to web3 dev stack product-market fit?

I read this as:

> How sure are we that this project will successfully achieve the
"impact" specified in the impact section?

However, I don't believe this was the intended question. The intention
was to determine how confident we are that we're solving the _right_
problem, not how confident we are that the proposal will solve the
problem.

I've changed the template to ask:

> How sure are we that this is the right problem to work on?